update: 인증 필요 여부에 따라 게시글 service 분리
